requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.16
aiosignal==1.3.1
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
asyncer==0.0.7
attrs==24.1.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.7.4
cffi==1.16.0
chainlit==2.4.400
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==1.0.5
chromedriver==2.24.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib @ file:///private/tmp/python-distlib-20231212-4938-trhlvy/distlib-0.3.8
distro==1.9.0
durationpy==0.9
et_xmlfile==2.0.0
executing==2.0.1
fake-useragent==2.0.3
fastapi==0.115.9
fastjsonschema==2.20.0
filelock @ file:///private/tmp/python-filelock-20231112-5622-k62fq8/filelock-3.13.1/dist/filelock-3.13.1-py3-none-any.whl#sha256=57dbda9b35157b05fb3e58ee91448612eb674172fab98ee235ccb0b5bee19a1c
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.7
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.3.23
langchain-community==0.0.38
langchain-core==0.3.55
langchain-openai==0.2.8
langchain-text-splitters==0.3.8
langsmith==0.1.143
Lazify==0.4.0
literalai==0.1.201
lxml==5.2.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mistune==3.0.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.35.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.75.0
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.0
opentelemetry-instrumentation-anthropic==0.39.0
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-bedrock==0.39.0
opentelemetry-instrumentation-chromadb==0.39.0
opentelemetry-instrumentation-cohere==0.39.0
opentelemetry-instrumentation-crewai==0.39.0
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-google-generativeai==0.39.0
opentelemetry-instrumentation-groq==0.39.0
opentelemetry-instrumentation-haystack==0.39.0
opentelemetry-instrumentation-lancedb==0.39.0
opentelemetry-instrumentation-langchain==0.39.0
opentelemetry-instrumentation-llamaindex==0.39.0
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.0
opentelemetry-instrumentation-milvus==0.39.0
opentelemetry-instrumentation-mistralai==0.39.0
opentelemetry-instrumentation-ollama==0.39.0
opentelemetry-instrumentation-openai==0.39.0
opentelemetry-instrumentation-pinecone==0.39.0
opentelemetry-instrumentation-qdrant==0.39.0
opentelemetry-instrumentation-replicate==0.39.0
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.0
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.0
opentelemetry-instrumentation-transformers==0.39.0
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.0
opentelemetry-instrumentation-watsonx==0.39.0
opentelemetry-instrumentation-weaviate==0.39.0
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
optuna==4.2.1
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
playwright==1.50.0
posthog==3.24.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==5.29.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.8.1
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.1.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-json-logger==2.0.7
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.19.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
selenium-stealth==1.0.6
Send2Trash==1.8.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.35
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.41.3
stevedore==5.1.0
streamlit==1.44.1
sympy==1.13.1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.6.0
tornado==6.4.1
tqdm==4.67.0
traceloop-sdk==0.39.0
traitlets==5.14.3
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
uptrace==1.31.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.34.1
uvloop==0.21.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==24.6.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
widgetsnbextension==4.0.11
wrapt==1.17.2
wsproto==1.2.0
yarl==1.17.1
zipp==3.21.0
